backend/requirements.txt
flask
pandas
numpy
python-dotenv
pymongo
statsmodels
pytz
scikit-learn
requests
google-genai
